requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
lxml==4.9.3
fake-useragent==1.4.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
psutil==5.9.6
aioredis==2.0.1
